Create a new portfolio by name
- connected button with a new portfolio dialog
- refresh navigation on the left on a new portfolio successful creation
