+++ tests/handlers/test_portfolio_creation_handler.py
@@ -1,10 +1,7 @@
-from datetime import datetime
-
 import pytest
 
 from ftt.handlers.portfolio_creation_handler import PortfolioCreationHandler
-from ftt.storage.models.portfolio import Portfolio
-from ftt.storage.models.portfolio_version import PortfolioVersion
+from ftt.storage.models import Portfolio
 
 
 class TestPortfolioCreationHandler:
@@ -14,24 +11,11 @@ class TestPortfolioCreationHandler:
 
     @pytest.fixture
     def arguments(self):
-        return {
-            "name": "Repository 1",
-            "value": 10000,
-            "period_start": datetime(2021, 1, 1),
-            "period_end": datetime(2021, 4, 25),
-            "interval": "1wk",
-        }
+        return {"name": "Portfolio 123"}
 
-    def test_creates_portfolio(self, subject, arguments):
+    def test_creates_portfolio_with_given_name(self, subject, arguments):
         result = subject.handle(**arguments)
 
         assert result.is_ok()
         assert type(result.value) == Portfolio
         assert result.value.id is not None
-
-    def test_creates_portfolio_first_version(self, subject, arguments):
-        result = subject.handle(**arguments)
-
-        assert result.is_ok()
-        assert type(result.value.versions[0]) == PortfolioVersion
-        assert result.value.versions[0].id is not None
